+++ src/Temporalio/Client/TemporalClient.Workflow.cs
@@ -335,82 +335,6 @@ namespace Temporalio.Client
                     req.NextPageToken = resp.NextPageToken;
                 }
             }
-
-            /// <inheritdoc />
-            public override async Task UpdateWorkerBuildIdCompatibilityAsync(UpdateWorkerBuildIdCompatibilityInput input)
-            {
-                var req = new UpdateWorkerBuildIdCompatibilityRequest
-                {
-                    Namespace = Client.Options.Namespace,
-                    TaskQueue = input.TaskQueue,
-                };
-                switch (input.BuildIdOp)
-                {
-                    case BuildIdOp.AddNewDefault op:
-                        req.AddNewBuildIdInNewDefaultSet = op.BuildId;
-                        break;
-                    case BuildIdOp.AddNewCompatible op:
-                        req.AddNewCompatibleBuildId =
-                        new UpdateWorkerBuildIdCompatibilityRequest.Types.AddNewCompatibleVersion
-                        {
-                            NewBuildId = op.BuildId,
-                            ExistingCompatibleBuildId = op.ExistingCompatibleBuildId,
-                            MakeSetDefault = op.MakeSetDefault,
-                        };
-                        break;
-                    case BuildIdOp.PromoteSetByBuildId op:
-                        req.PromoteSetByBuildId = op.BuildId;
-                        break;
-                    case BuildIdOp.PromoteBuildIdWithinSet op:
-                        req.PromoteBuildIdWithinSet = op.BuildId;
-                        break;
-                    case BuildIdOp.MergeSets op:
-                        req.MergeSets = new UpdateWorkerBuildIdCompatibilityRequest.Types.MergeSets
-                        {
-                            PrimarySetBuildId = op.PrimaryBuildId,
-                            SecondarySetBuildId = op.SecondaryBuildId,
-                        };
-                        break;
-                }
-
-                await Client.Connection.WorkflowService
-                    .UpdateWorkerBuildIdCompatibilityAsync(req, DefaultRetryOptions(input.RpcOptions))
-                    .ConfigureAwait(false);
-            }
-
-            /// <inheritdoc />
-            public override async Task<WorkerBuildIdVersionSets> GetWorkerBuildIdCompatibilityAsync(
-                GetWorkerBuildIdCompatibilityInput input)
-            {
-                var req = new GetWorkerBuildIdCompatibilityRequest
-                {
-                    Namespace = Client.Options.Namespace,
-                    TaskQueue = input.TaskQueue,
-                    MaxSets = input.MaxSets,
-                };
-                var resp = await Client.Connection.WorkflowService
-                    .GetWorkerBuildIdCompatibilityAsync(req, DefaultRetryOptions(input.RpcOptions))
-                    .ConfigureAwait(false);
-                return WorkerBuildIdVersionSets.FromProto(resp);
-            }
-
-            /// <inheritdoc />
-            public override async Task<WorkerTaskReachability> GetWorkerTaskReachabilityAsync(
-                GetWorkerTaskReachabilityInput input)
-            {
-                var req = new GetWorkerTaskReachabilityRequest
-                {
-                    Namespace = Client.Options.Namespace,
-                    Reachability = input.ReachabilityType?.ToProto() ?? TaskReachability.Unspecified,
-                };
-                req.BuildIds.AddRange(input.BuildIds);
-                req.TaskQueues.AddRange(input.TaskQueues);
-                var resp = await Client.Connection.WorkflowService
-                    .GetWorkerTaskReachabilityAsync(req, DefaultRetryOptions(input.RpcOptions))
-                    .ConfigureAwait(false);
-                return WorkerTaskReachability.FromProto(resp);
-            }
-
 #if NETCOREAPP3_0_OR_GREATER
             /// <inheritdoc />
             public override IAsyncEnumerable<WorkflowExecution> ListWorkflowsAsync(
