+++ tests/Temporalio.Tests/Client/TemporalClientTaskQueueTests.cs
@@ -18,38 +18,38 @@ public class TemporalClientTaskQueueTests : WorkflowEnvironmentTestBase
 
         await Client.UpdateWorkerBuildIdCompatibilityAsync(taskQueue, new BuildIdOp.AddNewDefault("1.0"));
 
-        var sets = await Client.GetWorkerBuildIdCompatibilityAsync(taskQueue);
+        var sets = (await Client.GetWorkerBuildIdCompatibilityAsync(taskQueue))!;
         Assert.Equal(1, sets.VersionSets.Count);
-        Assert.Equal("1.0", sets.DefaultBuildId());
+        Assert.Equal("1.0", sets.DefaultBuildId);
 
         await Client.UpdateWorkerBuildIdCompatibilityAsync(taskQueue, new BuildIdOp.AddNewCompatible("1.1", "1.0"));
 
-        sets = await Client.GetWorkerBuildIdCompatibilityAsync(taskQueue);
+        sets = (await Client.GetWorkerBuildIdCompatibilityAsync(taskQueue))!;
         Assert.Equal(1, sets.VersionSets.Count);
-        Assert.Equal("1.1", sets.DefaultBuildId());
+        Assert.Equal("1.1", sets.DefaultBuildId);
 
         await Client.UpdateWorkerBuildIdCompatibilityAsync(taskQueue, new BuildIdOp.PromoteBuildIdWithinSet("1.0"));
 
-        sets = await Client.GetWorkerBuildIdCompatibilityAsync(taskQueue);
+        sets = (await Client.GetWorkerBuildIdCompatibilityAsync(taskQueue))!;
         Assert.Equal(1, sets.VersionSets.Count);
-        Assert.Equal("1.0", sets.DefaultBuildId());
+        Assert.Equal("1.0", sets.DefaultBuildId);
 
         await Client.UpdateWorkerBuildIdCompatibilityAsync(taskQueue, new BuildIdOp.AddNewDefault("2.0"));
 
-        sets = await Client.GetWorkerBuildIdCompatibilityAsync(taskQueue);
+        sets = (await Client.GetWorkerBuildIdCompatibilityAsync(taskQueue))!;
         Assert.Equal(2, sets.VersionSets.Count);
-        Assert.Equal("2.0", sets.DefaultBuildId());
+        Assert.Equal("2.0", sets.DefaultBuildId);
 
         await Client.UpdateWorkerBuildIdCompatibilityAsync(taskQueue, new BuildIdOp.PromoteSetByBuildId("1.0"));
 
-        sets = await Client.GetWorkerBuildIdCompatibilityAsync(taskQueue);
+        sets = (await Client.GetWorkerBuildIdCompatibilityAsync(taskQueue))!;
         Assert.Equal(2, sets.VersionSets.Count);
-        Assert.Equal("1.0", sets.DefaultBuildId());
+        Assert.Equal("1.0", sets.DefaultBuildId);
 
         await Client.UpdateWorkerBuildIdCompatibilityAsync(taskQueue, new BuildIdOp.MergeSets("2.0", "1.0"));
 
-        sets = await Client.GetWorkerBuildIdCompatibilityAsync(taskQueue);
+        sets = (await Client.GetWorkerBuildIdCompatibilityAsync(taskQueue))!;
         Assert.Equal(1, sets.VersionSets.Count);
-        Assert.Equal("2.0", sets.DefaultBuildId());
+        Assert.Equal("2.0", sets.DefaultBuildId);
     }
 }
